Ajout méthode gérer, add, sup, voir, toString, voirLivre, afficherAd, sauvegarder, reecrire et chargerFicher pour les adhérents

+++ Bibliothèque/src/Main.java
@@ -9,15 +9,15 @@ import state.*;
 public class Main {
     public static void main(String[] args) {
     	Bibliotheque biblio = new Bibliotheque();
+    	biblio.chargerAdherentsDepuisFichier();
     	biblio.chargerLivresDepuisFichier();
     	
-        Utilisateur admin = UtilisateurFactory.creerUtilisateur("admin", "Alice", "alice@bib.com");
-        Utilisateur adherent = UtilisateurFactory.creerUtilisateur("adherent", "Bob", "bob@bib.com");
+    	AuthentificationService auth = new AuthentificationService();
 
         Scanner scanner = new Scanner(System.in);
 
         while (true) {
-            Utilisateur utilisateur = connexion(scanner);
+            Utilisateur utilisateur = connexion(scanner, auth);
             boolean active = true;
             
             while (active) {
@@ -27,33 +27,56 @@ public class Main {
                 int choix = scanner.nextInt();
                 scanner.nextLine();
 
-                switch (choix) {
-                    case 1:
-                    	((Admin) utilisateur).ajouterLivre(scanner, biblio); //Ajouter un livre
-                        break;
-                    case 2:
-                    	((Admin) utilisateur).supprimerLivre(scanner, biblio); //Supprimer un livre
-                        break;
-                    case 3:
-                    	((Admin) utilisateur).gererAdherent(scanner, biblio); //Gérer les adhérents
-                        break;
-                    case 4:
-                    	((Admin) utilisateur).voirBiblio(scanner, biblio); //Voir les livres dans la bibliothèque
-                        break;
-                    case 5:
-                        System.out.println("Déconnexion");
-                        active = false;
-                        break;
-                    default:
-                        System.out.println("Choix invalide.");
-                        break;
+                //Switch pour les commandes de l'admin
+                if(utilisateur instanceof Admin) {
+	                switch (choix) {
+	                    case 1:
+	                    	((Admin) utilisateur).ajouterLivre(scanner, biblio); //Ajouter un livre
+	                        break;
+	                    case 2:
+	                    	((Admin) utilisateur).supprimerLivre(scanner, biblio); //Supprimer un livre
+	                        break;
+	                    case 3:
+	                    	((Admin) utilisateur).gererAdherent(scanner, biblio, auth); //Gérer les adhérents
+	                        break;
+	                    case 4:
+	                    	((Admin) utilisateur).voirBiblio(scanner, biblio); //Voir les livres dans la bibliothèque
+	                        break;
+	                    case 5:
+	                        System.out.println("Déconnexion");
+	                        active = false;
+	                        break;
+	                    default:
+	                        System.out.println("Choix invalide.");
+	                        break;
+	                }
+                }else if(utilisateur instanceof Adherent){
+                
+	                //Switch pour les commandes de l'adhérent
+	                switch (choix) {
+	                case 1:
+	                	((Adherent) utilisateur).voirLivre(scanner, biblio); //Voir les livres dans la bibliothèque
+	                    break;
+	                case 2:
+	                    break;
+	                case 3:
+	                    break;
+	                case 4:
+	                    break;
+	                case 5:
+	                    System.out.println("Déconnexion");
+	                    active = false;
+	                    break;
+	                default:
+	                    System.out.println("Choix invalide.");
+	                    break;
+	                }
                 }
             }
         }
     }
 
-    public static Utilisateur connexion(Scanner scanner) {
-        AuthentificationService auth = new AuthentificationService();
+    public static Utilisateur connexion(Scanner scanner, AuthentificationService auth) {
         Utilisateur utilisateur = null;
 
         while (utilisateur == null) {
